project2/requirements.txt
Django==2.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
